+++ tensorflow/core/common_runtime/pluggable_device/BUILD
@@ -72,7 +72,7 @@ cc_library(
     visibility = ["//visibility:public"],
     deps = [
         "//tensorflow/c/experimental/grappler",
-        "//tensorflow/c/experimental/profiler",
+        "//tensorflow/c/experimental/pluggable_profiler",
         "//tensorflow/c/experimental/stream_executor",
         "//tensorflow/c/experimental/stream_executor:stream_executor_internal",
         "//tensorflow/core:core_cpu",
